refactor: use named constants for dict and segment arena related offset in hints
