+++ src/vm/virtualMachine.test.ts
@@ -1,11 +1,7 @@
 import { test, expect, describe, spyOn } from 'bun:test';
 
 import { ExpectedFelt, ExpectedRelocatable } from 'errors/primitives';
-import {
-  DictNotFound,
-  InvalidBufferResOp,
-  UnusedRes,
-} from 'errors/virtualMachine';
+import { InvalidBufferResOp, UnusedRes } from 'errors/virtualMachine';
 
 import { Felt } from 'primitives/felt';
 import { Relocatable } from 'primitives/relocatable';
@@ -19,7 +15,7 @@ import {
   FpUpdate,
   Op1Src,
 } from './instruction';
-import { Dictionary, VirtualMachine } from './virtualMachine';
+import { VirtualMachine } from './virtualMachine';
 import { CellRef, Operation, OpType, ResOperand } from 'hints/hintParamsSchema';
 
 const instructions = {
@@ -739,26 +735,6 @@ describe('VirtualMachine', () => {
           expect(vm.getResOperandValue(resOperand)).toEqual(expected);
         }
       );
-    });
-  });
-
-  describe('Dictionary', () => {
-    test('should properly initialize the dict manager', () => {
-      const vm = new VirtualMachine();
-      expect(vm.dictManager.size).toEqual(0);
-    });
-
-    test('should properly create a new dictionary', () => {
-      const vm = new VirtualMachine();
-      const address = vm.newDict();
-      expect(address).toEqual(new Relocatable(0, 0));
-      expect(vm.getDict(address)).toEqual(new Dictionary(new Felt(0n)));
-    });
-
-    test('should throw DictNotFound when accessing a non-existing dictionary', () => {
-      const vm = new VirtualMachine();
-      const address = new Relocatable(2, 3);
-      expect(() => vm.getDict(address)).toThrow(new DictNotFound(address));
     });
   });
 });
